+++ include/Game/Point.h
@@ -11,8 +11,8 @@
  * @brief Represents a point in 2D space with x and y coordinates.
  */
 struct Point {
-    int x; /**< The x-coordinate of the point. */
-    int y; /**< The y-coordinate of the point. */
+    float x; /**< The x-coordinate of the point. */
+    float y; /**< The y-coordinate of the point. */
 };
 
 #endif //PLAY_TOGETHER_POINT_H
